detach movement from core gaming module
