+++ Altruist/Gaming/Main/Core/General/Models.cs
@@ -531,43 +531,6 @@ public abstract class PlayerEntity : VaultModel, ISynchronizedEntity
     }
 }
 
-public abstract class MovementInput
-{
-    public float RotationSpeed { get; set; } = 0f;
-    public bool Turbo { get; set; }
-}
-
-
-public class ForwardMovementInput : MovementInput
-{
-    public bool MoveUp { get; set; }
-    public bool RotateLeft { get; set; }
-    public bool RotateRight { get; set; }
-}
-
-
-public class EightDirectionMovementInput : MovementInput
-{
-    public bool MoveUp { get; set; }
-    public bool MoveDown { get; set; }
-    public bool MoveLeft { get; set; }
-    public bool MoveRight { get; set; }
-}
-
-
-public class VehicleMovementInput : EightDirectionMovementInput
-{
-
-}
-
-
-public class SpaceshipMovementInput : ForwardMovementInput
-{
-
-}
-
-
-
 public abstract class Vehicle : PlayerEntity
 {
     [VaultColumn]
