+++ Altruist/Gaming/Main/Core/Shooter/ShooterPortal.cs
@@ -18,7 +18,6 @@ using Microsoft.Extensions.Logging;
 
 namespace Altruist.Gaming.Shooter;
 
-
 public abstract class AltruistShootingPortal<TPlayerEntity> : AltruistGameSessionPortal<TPlayerEntity> where TPlayerEntity : PlayerEntity, new()
 {
     public AltruistShootingPortal(PortalContext context, GameWorldCoordinator gameWorldCoordinator, IPlayerService<TPlayerEntity> playerService, ILoggerFactory loggerFactory)
