+++ Altruist/Gaming/Movement/Core/Contracts/Service.cs
@@ -1,0 +1,6 @@
+namespace Altruist.Gaming.Movement;
+
+public interface IMovementService<TPlayerEntity> where TPlayerEntity : PlayerEntity
+{
+    Task<TPlayerEntity?> MovePlayerAsync(string playerId, IMovementPacket input);
+}
